requirements.txt
Flask==3.1.3
PyJWT==2.10.1
blinker==1.9.0
click==8.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.6
flask-cors
flask-sqlalchemy
psycopg2-binary